Refactored file structure for consistency.
