+++ src/index.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
+
 import "./styles/index.css";
 import App from "./App";
 import BadgeProvider from "./context/BadgeContext";
