+++ src/pages/AllStampsEarned/styles.tsx
@@ -1,0 +1,5 @@
+const styles = {
+  container: "App overflow-scroll bg-background h-screen",
+};
+
+export default styles;
