+++ src/pages/BadgeModal/BadgeAnimation.css
