+++ src/pages/BadgeModal/styles.tsx
