+++ src/pages/MapPage/MapPage.tsx
@@ -1,19 +1,15 @@
 import { useRef } from "react";
 
-import BadgeMap from "../components/BadgeMap/BadgeMap";
-import InfoBox from "../components/InfoBox";
-import Button from "../components/Button";
-import { useBadge } from "../context/BadgeContext";
-import BadgeWidget from "../components/BadgeWidget/BadgeWidget";
+import styles from "./styles";
+import BadgeMap from "./components/BadgeMap/BadgeMap";
+import InfoBox from "../../components/InfoBox/InfoBox";
+import Button from "../../components/Buttons/Button/Button";
+import { useBadge } from "../../context/BadgeContext";
+import BadgeWidget from "./components/BadgeWidget/BadgeWidget";
 
 const MapPage = (): JSX.Element => {
   const { getFoundBadges } = useBadge();
   const foundBadges = useRef(getFoundBadges());
-  const styles = {
-    mainContainer:
-      "App px-10 py-8 overflow-scroll bg-background flex-col h-screen",
-    headerText: "text-center text-3xl text-darker font-semibold mb-5",
-  };
 
   return (
     <div className={styles.mainContainer}>
