+++ src/pages/MapPage/components/BadgeMap/styles.tsx
