+++ src/pages/MapPage/components/BadgeMap/utilities/MapImageData.tsx
