+++ src/pages/MapPage/components/BadgeMap/utilities/PinData.tsx
