+++ src/pages/MapPage/components/BadgeWidget/styles.tsx
