+++ src/pages/MapPage/styles.tsx
@@ -1,0 +1,7 @@
+const styles = {
+  mainContainer:
+    "App px-10 py-8 overflow-scroll bg-background flex-col h-screen",
+  headerText: "text-center text-3xl text-darker font-semibold mb-5",
+};
+
+export default styles;
