+++ src/pages/OnboardPage/components/styles.tsx
