+++ src/styles/index.css
@@ -14,6 +14,10 @@ body {
   overflow: hidden;
 }
 
+h1 {
+  font-size: 20px;
+}
+
 body > * {
   /* display: flex; */
 }
